pyproject.toml
[tool.poetry]
name = "blog-sn"
version = "0.1.0"
description = ""
authors = ["Maksym Sydorchuk <maximsidorchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
fastapi = {extras = ["standard"], version = "0.115.2"}
pydantic-settings = "2.6.0"
sqlalchemy = "2.0.36"
python-jose = "3.3.0"
passlib = "1.7.4"
asyncpg = "0.30.0"
alembic = "1.13.3"
greenlet = "3.1.1"
openai = "1.52.1"
celery = "5.4.0"
fastapi-cache2 = {extras = ["redis"], version = "0.2.2"}
flower = "2.0.1"
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
httpx = "0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
